+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.12.0 (2024-01-04)
+
+* Add `StreamOpts::interrupted_next_item_include` to toggle whether interrupted item is streamed.
+* Separate `FnGraph::stream` and `FnGraph::stream_with`, from `FnGraph::stream_interruptible` and `FnGraph::stream_with_interruptible`.
+
+
 ## 0.11.0 (2024-01-03)
 
 * Fix `StreamOutcome` `fn_ids_processed` and `fn_ids_not_processed` values.